requirements.txt
PyJWT==1.7.1
requests==2.27.1
passlib==1.7.2

psycopg2-binary==2.9.6
python-multipart==0.0.5
boto3==1.17.27
alembic==1.8.0
fastapi-sqlalchemy==0.2.1
SQLAlchemy==2.0.13
mailchimp-marketing==3.0.80
git+https://github.com/mailchimp/mailchimp-transactional-python.git
stripe==2.53.0
Pillow==8.4.0
fastapi>=0.103.2
pydantic>=1.8.0,<2.0.0
uvicorn[standard]
git+https://github.com/Packstack-Tech/models.git@v1.1.1#egg=models